Refactored to cache class
