+++ src/main/kotlin/no/nav/syfo/client/person/adressebeskyttelse/AdressebeskyttelseClient.kt
@@ -17,17 +17,13 @@ import no.nav.syfo.domain.PersonIdentNumber
 import no.nav.syfo.util.*
 import org.slf4j.LoggerFactory
 import redis.clients.jedis.Jedis
-import redis.clients.jedis.params.SetParams
 
 class AdressebeskyttelseClient(
-    private val jedis: Jedis,
+    jedis: Jedis,
     syfopersonBaseUrl: String
 ) {
-    private val personAdressebeskyttelseUrl: String
-
-    init {
-        this.personAdressebeskyttelseUrl = "$syfopersonBaseUrl$PERSON_ADRESSEBESKYTTELSE_PATH"
-    }
+    private val personAdressebeskyttelseUrl: String = "$syfopersonBaseUrl$PERSON_ADRESSEBESKYTTELSE_PATH"
+    private val cache = AdressebeskyttelseCache(jedis = jedis);
 
     private val httpClient = HttpClient(CIO) {
         install(JsonFeature) {
@@ -43,14 +39,7 @@ class AdressebeskyttelseClient(
         token: String,
         callId: String
     ): Boolean {
-        val cacheKey = cacheKey(personIdentNumber)
-        jedis.use { jedisClient ->
-            jedisClient.get(cacheKey)?.let { value ->
-                log.info("JTRACE: return cached value $value")
-                return value.toBoolean()
-            }
-        }
-        return try {
+        return cache.hasAdressebeskyttelse(personIdentNumber) ?: try {
             val response: HttpResponse = httpClient.get(personAdressebeskyttelseUrl) {
                 header(HttpHeaders.Authorization, bearerHeader(token))
                 header(NAV_CALL_ID_HEADER, callId)
@@ -59,14 +48,7 @@ class AdressebeskyttelseClient(
             }
             val adressebeskyttelseResponse = response.receive<AdressebeskyttelseResponse>()
             COUNT_CALL_PERSON_ADRESSEBESKYTTELSE_SUCCESS.inc()
-            jedis.use { jedisClient ->
-                jedisClient.set(
-                    cacheKey,
-                    adressebeskyttelseResponse.beskyttet.toString(),
-                    SetParams().ex(CACHE_ADRESSEBESKYTTELSE_EXPIRE_SECONDS)
-                )
-            }
-            log.info("JTRACE: set cached value ${adressebeskyttelseResponse.beskyttet}")
+            cache.setAdressebeskyttelse(personIdentNumber, adressebeskyttelseResponse.beskyttet)
             adressebeskyttelseResponse.beskyttet
         } catch (e: ClientRequestException) {
             handleUnexpectedReponseException(e.response)
@@ -84,16 +66,9 @@ class AdressebeskyttelseClient(
         return true
     }
 
-    private fun cacheKey(personIdentNumber: PersonIdentNumber): String {
-        return "$CACHE_ADRESSEBESKYTTELSE_KEY_PREFIX${personIdentNumber.value}"
-    }
-
     companion object {
         const val PERSON_PATH = "/syfoperson/api/person"
         const val PERSON_ADRESSEBESKYTTELSE_PATH = "$PERSON_PATH/adressebeskyttelse"
-
-        const val CACHE_ADRESSEBESKYTTELSE_KEY_PREFIX = "person-adressebeskyttelse-"
-        const val CACHE_ADRESSEBESKYTTELSE_EXPIRE_SECONDS = 3600
 
         private val log = LoggerFactory.getLogger(AdressebeskyttelseClient::class.java)
     }
